Try to make travis happy: next try

+++ vtm-tests/test/org/oscim/utils/ColorTest.java
@@ -18,6 +18,8 @@ public class ColorTest {
     public static final String ANSI_PREFIX_BACKGROUND_24Bit = "\033[48;2;";
     public static final String ANSI_PREFIX_FOREGROUND_24Bit = "\033[38;2;";
 
+    public static final boolean DISPLAY_IN_BROWSER = false;
+
     @Test
     public void testColorHSV() {
         List<Integer> colors = new ArrayList<>();
@@ -55,33 +57,37 @@ public class ColorTest {
 
     private void printColors(List<Integer> colors) {
 
-        try {
-            File tempFile;
-            tempFile = File.createTempFile("test-color-", ".html");
-            tempFile.deleteOnExit();
-            StringBuilder builder = new StringBuilder("<html>");
+        for (int color : colors) {
+            // Try to display them in terminal (Intellij not supports 24 bit colors)
+            System.out.println(ANSI_PREFIX_BACKGROUND_24Bit + Color.r(color) + ";" + Color.g(color) + ";" + Color.b(color) + "m " + Integer.toString(color) + ANSI_RESET);
+        }
 
-            for (int color : colors) {
-                // Try to display them in terminal (Intellij not supports 24 bit colors)
-                System.out.println(ANSI_PREFIX_BACKGROUND_24Bit + Color.r(color) + ";" + Color.g(color) + ";" + Color.b(color) + "m " + Integer.toString(color) + ANSI_RESET);
+        if (DISPLAY_IN_BROWSER) {
+            try {
+                File tempFile;
+                tempFile = File.createTempFile("test-color-", ".html");
+                tempFile.deleteOnExit();
+                StringBuilder builder = new StringBuilder("<html>");
 
-                builder.append(String.format("<div style=\"background:rgb(%s,%s,%s);\">", Color.r(color), Color.g(color), Color.b(color)));
-                builder.append(Color.toString(color));
-                builder.append("</div>");
+                for (int color : colors) {
+                    builder.append(String.format("<div style=\"background:rgb(%s,%s,%s);\">", Color.r(color), Color.g(color), Color.b(color)));
+                    builder.append(Color.toString(color));
+                    builder.append("</div>");
+                }
+                builder.append("</html>");
+
+                BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
+                writer.write(builder.toString());
+                writer.close();
+
+                Desktop.getDesktop().browse(tempFile.toURI());
+
+                Thread.sleep(2000);
+            } catch (IOException e) {
+                e.printStackTrace();
+            } catch (InterruptedException e) {
+                e.printStackTrace();
             }
-            builder.append("</html>");
-
-            BufferedWriter writer = new BufferedWriter(new FileWriter(tempFile));
-            writer.write(builder.toString());
-            writer.close();
-
-            Desktop.getDesktop().browse(tempFile.toURI());
-
-            Thread.sleep(2000);
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
         }
     }
 }
